fix measure of travelled distance
We used to project the vector of translation between original and final pose on the Y axis of the initial frame. However that frame is not guaranteed to be horizontal, specifically when the Xtion is tilted or the robot's position is weird. Consequently, we project on the XZ plane the previously mentioned vector.